requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
rapidfuzz==3.10.0
pyyaml==6.0.2
python-dotenv==1.0.1
